requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
cffi==1.14.1
Django==3.1
django-registration-redux==2.8
Pillow==7.2.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
